Conflicts Resolved:
	release/windows/msix/AppxManifest.xml.template
	scripts/startup/bl_ui/node_add_menu_geometry.py
	scripts/startup/bl_ui/properties_render.py
	scripts/startup/bl_ui/space_userpref.py
	source/blender/asset_system/intern/library_types/essentials_library.cc
	source/blender/blenkernel/intern/appdir.cc
	source/blender/editors/interface/views/tree_view.cc
	source/blender/editors/space_clip/clip_buttons.cc
	source/blender/editors/space_node/drawnode.cc
	source/blender/editors/space_view3d/view3d_buttons.cc
	source/blender/editors/uvedit/uvedit_buttons.cc
	source/blender/makesdna/DNA_scene_types.h
	source/blender/makesdna/DNA_screen_types.h
	source/blender/makesdna/DNA_userdef_types.h
	source/blender/makesdna/DNA_view3d_types.h
	source/blender/makesrna/intern/rna_node_tree_interface.cc
	source/blender/makesrna/intern/rna_nodetree.cc

+++ source/blender/render/intern/render_result.h
@@ -117,7 +117,7 @@ struct ImBuf *render_result_rect_to_ibuf(struct RenderResult *rr,
 
 void render_result_rect_fill_zero(struct RenderResult *rr, int view_id);
 void render_result_rect_get_pixels(struct RenderResult *rr,
-                                   unsigned int *rect,
+                                   uint8_t *rect,
                                    int rectx,
                                    int recty,
                                    const struct ColorManagedViewSettings *view_settings,
